+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/TeacherRestController.java
@@ -1,6 +1,9 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
 
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.TeacherDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.TeacherController;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Teacher;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -13,5 +16,10 @@ public class TeacherRestController {
 
     public TeacherRestController(TeacherController teacherController) {
         this.teacherController = teacherController;
+    }
+
+    @PostMapping("")
+    public Teacher createTeacher(@RequestBody TeacherDTO teacherDTO){
+        return teacherController.createTeacher(teacherDTO);
     }
 }