+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/TeacherController.java
@@ -1,5 +1,6 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers;
 
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.TeacherDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Teacher;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.service.TeacherService;
 import org.springframework.stereotype.Controller;
@@ -13,4 +14,8 @@ public class TeacherController {
         this.teacherService = teacherService;
     }
 
+    public Teacher createTeacher(TeacherDTO teacherDTO) {
+        Teacher teacher = new Teacher(teacherDTO);
+        return teacherService.createTeacher(teacher);
+    }
 }