+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/TeacherService.java
@@ -1,5 +1,6 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.service;
 
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Teacher;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.persistence.TeacherRepository;
 import org.springframework.stereotype.Service;
 
@@ -10,5 +11,9 @@ public class TeacherService {
 
     public TeacherService(TeacherRepository teacherRepository) {
         this.teacherRepository = teacherRepository;
+    }
+
+    public Teacher createTeacher(Teacher teacher) {
+        return teacherRepository.save(teacher);
     }
 }